update/delete of controller/service for category

+++ src/category/category.service.ts
@@ -4,7 +4,7 @@ import { Prisma } from '@prisma/client';
 
 @Injectable()
 export class CategoryService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private prisma: PrismaService) { }
   async exists(where: Prisma.CategoryWhereUniqueInput) {
     return this.prisma.category.findFirst({ where });
   }
@@ -15,6 +15,8 @@ export class CategoryService {
   }
 
   async deleteCategory(where: Prisma.CategoryWhereUniqueInput) {
+    const IsExist = await this.exists(where);
+    if (!IsExist) throw new BadRequestException('Category does not exists');
     await this.prisma.category.delete({ where });
   }
 
@@ -26,7 +28,9 @@ export class CategoryService {
     where: Prisma.CategoryWhereUniqueInput,
     data: Prisma.CategoryUpdateInput,
   ) {
-    await this.prisma.category.update({
+    const IsExist = await this.exists(where);
+    if (!IsExist) throw new BadRequestException('Category does not exists');
+    return await this.prisma.category.update({
       data,
       where,
     });
